Update README.md
Steps to Access the Site's Content:

1. Enable JavaScript in Your Browser:

Google Chrome:

Go to Settings > Privacy and security > Site Settings > JavaScript.

Ensure that JavaScript is allowed.


Mozilla Firefox:

Type about:config in the address bar.

Search for javascript.enabled and ensure it's set to true.


Microsoft Edge:

Go to Settings > Cookies and site permissions > JavaScript.

Ensure that JavaScript is allowed.


Safari:

Go to Preferences > Security.

Check the box next to Enable JavaScript.




2. Use a Modern Browser:

Ensure you're using an up-to-date browser that supports JavaScript, such as the latest versions of Chrome, Firefox, Edge, or Safari.



3. Access the Site After Enabling JavaScript:

Once JavaScript is enabled, revisit https://groupg.wuaze.com.

The site's content, including any documents or README files, should now be accessible.




Alternative Approach:

If you're still unable to access the site's content after enabling JavaScript, consider the following:

Contact the Site Administrator:

If you have contact information for the site's administrator or the individual who shared the link with you, reach out to them for assistance or to request the documents directly.


Check for a Public Repository:

If the site is associated with a GitHub repository or another public code repository, you might find the README and other documentation there.



Understanding README Files:

A README file is a text file that provides information about a project.  It's typically the first file a visitor will see when visiting a repository.  A well-structured README helps users understand the purpose of the project, how to set it up, and how to use it. 

Common Sections in a README File:

Project Title: The name of the project.

Description: A brief overview of what the project does.

Installation:To run the groupg.wuaze.com website locally on your machine, follow these easy steps:
Prerequisites
A modern web browser (Chrome, Firefox, Edge, Safari)


Git (optional, if you want to clone the project repository)


Steps
Clone the repository
Navigate into the project folder:

Usage:open  the link in any browser

License: Information about the project's license. ## License

This project, **groupg.wuaze.com**, is licensed under the Site.pro License.  
See the link for full details.


### About the Site.proLicense

The Site.pro License is a permissive open-source license that allows you to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of this software, provided that the original copyright notice and this permission notice are included in all copies or substantial portions of the software.

By using this license, **groupg.wuaze.com** encourages open collaboration while protecting contributors’ rights.

© 2025 GroupG Team